--- Auto-Git Commit ---

+++ backend/controllers/sauce.js
@@ -1,23 +1,20 @@
 const Sauce = require('../models/Sauce');
+const fs = require('fs');
 
 // Ajoute un objet Sauce
 exports.createSauce = (req, res, next) => {
-  Sauce.findOne({ _id: req.params.id })
-  .then(sauce => {
-    if (sauce.userId != req.auth.userId) {
-      res.status(401).json({ message: 'Not authorized' });
-    } else {
-      const filename = sauce.imageUrl.split('/images/')[1];
-      fs.unlink(`images/${filename}`, () => {
-        Sauce.deleteOne({ _id: req.params.id })
-          .then(() => res.status(200).json({ message: 'Sauce supprimée !' }))
-          .catch(error => res.status(401).json({ error }));
-      });
-    }
-  })
-  .catch(error => {
-    res.status(500).json({ error });
+  const sauceObject = JSON.parse(req.body.sauce);
+  delete sauceObject._id;
+  delete sauceObject._userId;
+  const sauce = new Sauce({
+    ...sauceObject,
+    userId: req.auth.userId,
+    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
   });
+
+  sauce.save()
+    .then(() => { res.status(201).json({ message: 'Sauce enregistrée !' }) })
+    .catch(error => { res.status(400).json({ error }) })
 };
 
 // Modification d'un objet Sauce
@@ -45,9 +42,22 @@ exports.modifySauce = (req, res, next) => {
 
 // Suppression objet Sauce
 exports.deleteSauce = (req, res, next) => {
-  Sauce.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json())
-    .catch(error => res.status(404).json({ error }));
+  Sauce.findOne({ _id: req.params.id })
+    .then(sauce => {
+      if (sauce.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Not authorized' });
+      } else {
+        const filename = sauce.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Sauce.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
+            .catch(error => res.status(401).json({ error }));
+        });
+      }
+    })
+    .catch(error => {
+      res.status(500).json({ error });
+    });
 }
 
 // Lecture d'un objet Sauce
